@@ -29,11 +29,22 @@ public class Controlador implements EventListener {
                     pv = privateChannel;
                     return privateChannel.sendMessage("número 1").submit();
                 });
-        log.info("aaa");
-        Executador.poolExecutor.submit(()->{
-            EventLoop.getInstance(event.getJDA());
+        Executador.poolExecutor.execute(()->{
+            boolean feonline = false;
+            while(true){
+                try {
+                    Thread.sleep(5000);
+                } catch (InterruptedException e) {e.printStackTrace();}
+                event.getJDA().getGuildById(286964489406906368l).retrieveMemberById(287269006753267713l).queue(member -> {
+
+                    System.out.println(member.getEffectiveName() + " :: " + member.getOnlineStatus());
+                });
+
+            }
         });
-        log.info("bbb");
+        Executador.poolExecutor.execute(()->
+            EventLoop.getInstance(event.getJDA())
+        );
     }
 
     public void onGuildMessageReceived(@NotNull GuildMessageReceivedEvent event) {
@@ -47,38 +58,41 @@ public class Controlador implements EventListener {
     }
 
     public void onGuildMessageReactionAdd(@NotNull GuildMessageReactionAddEvent event) {
-        if(event.getUser() != null) {
+        if(event.getMember() != null) {
             if (!event.getUser().isBot())
-                Executador.interpreta(event, event.getUser());
+                Executador.interpreta(event, event.getMember());
         }else
-            event.getJDA().retrieveUserById(event.getUserIdLong()).queue(u->{
-                if (!u.isBot())
+            event.getGuild().retrieveMemberById(event.getUserIdLong()).queue(u->{
+                if (!u.getUser().isBot()) {
+                    log.trace("chamanda para evento");
                     Executador.interpreta(event, u);
+                }
             });
    }
 
     public void onGuildMessageReactionRemove(@NotNull GuildMessageReactionRemoveEvent event) {
-        if(event.getUser() != null) {
-            if (!event.getUser().isBot())
-                Executador.interpreta(event, event.getUser());
+        if(event.getMember() != null) {
+            if (!event.getMember().getUser().isBot())
+                Executador.interpreta(event, event.getMember());
         }else
-            event.getJDA().retrieveUserById(event.getUserIdLong()).queue(u->{
-                if (!u.isBot())
+            event.getGuild().retrieveMemberById(event.getUserIdLong()).queue(u->{
+                if (!u.getUser().isBot())
                     Executador.interpreta(event, u);
             });
     }
 
     public void slashCommandEvent(@NotNull SlashCommandEvent event){
-        log.info("slah!!!");
+        log.debug("slah!!!");
         Executador.interpreta(event);
     }
 
     @Override
     public void onEvent(@NotNull GenericEvent event) {
-        if(event instanceof GuildMessageReceivedEvent) onGuildMessageReceived((GuildMessageReceivedEvent)event);
-        else if(event instanceof SlashCommandEvent) slashCommandEvent((SlashCommandEvent) event);
-        else if(event instanceof GuildMessageReactionAddEvent) onGuildMessageReactionAdd((GuildMessageReactionAddEvent)event);
-        else if(event instanceof GuildMessageReactionRemoveEvent) onGuildMessageReactionRemove((GuildMessageReactionRemoveEvent)event);
-        else if(event instanceof ReadyEvent) onReady((ReadyEvent)event);
+
+        if(event instanceof GuildMessageReceivedEvent e) onGuildMessageReceived(e);
+        else if(event instanceof SlashCommandEvent e) slashCommandEvent(e);
+        else if(event instanceof GuildMessageReactionAddEvent e) onGuildMessageReactionAdd(e);
+        else if(event instanceof GuildMessageReactionRemoveEvent e) onGuildMessageReactionRemove(e);
+        else if(event instanceof ReadyEvent e) onReady(e);
     }
 }
