@@ -3,6 +3,7 @@ package rode.comando.guild.mina;
 import net.dv8tion.jda.api.EmbedBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import rode.core.Anotacoes.EComandoJogo;
 import rode.model.ComandoGuild;
 import rode.core.EventLoop;
 import rode.core.Helper;
@@ -12,29 +13,22 @@ import rode.utilitarios.Constantes;
 
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.ResourceBundle;
 
+@EComandoJogo
 public class CriaCampoMinado extends ComandoGuild {
+    private static final Logger log = LoggerFactory.getLogger(CriaCampoMinado.class);
     public CriaCampoMinado() {
-        super(/*"mine"*/null, null, "campo");
+        super("campo", null, "campo");
     }
 
-    @Override
-    public void help(EmbedBuilder me, ResourceBundle loc) {
-
-    }
-
-    @Override
-    public void helpExtensive(EmbedBuilder me, ResourceBundle loc) {
-
-    }
 
     @Override
     public void execute(String[] args, Helper.Mensagem hm) throws Exception {
-        hm.reply(Constantes.builder(hm.bundle()),msg->{
-            hm.mensagem(msg);
+        hm.reply(Constantes.builder(hm.getBundle()), msg->{
+            hm.setMensagem(msg);
             EventLoop.addReacao(new ConversaCampo(hm));
+            log.trace("msg");
         });
     }
     public static class ConversaCampo extends MensagemReacao{
@@ -45,23 +39,23 @@ public class CriaCampoMinado extends ComandoGuild {
         private boolean check = false;
         private boolean end = false;
         public ConversaCampo(Helper hr) {
-            super(hr,hr.mensagem(), Arrays.asList(hr.id()), System.currentTimeMillis()+20000,null ,new HashMap<>());
-            delay(60000);
+            super(hr,hr.getMensagem(), Arrays.asList(hr.getId()), System.currentTimeMillis()+20000,null ,new HashMap<>());
+            setDelay(60000);
             this.campo = new CampoMinado();
-            src(new HashMap<>(){{
+            setComandos(new HashMap<>(){{
                 for(int i=0;i<10;i++) {
-                    final var ii = i;
+                    var ii = i;
                     put(Constantes.emote("" + i), r -> click(ii, r));
                 }
             }});
-            mensagem().clearReactions().queue(q->{
+            getMensagem().clearReactions().queue(q->{
                 for(int i=0;i<10;i++)
-                    mensagem().addReaction(Constantes.emote(""+i)).queue();
+                    getMensagem().addReaction(Constantes.emote(""+i)).queue();
             });
             eb = Constantes.builder()
                     .setTitle("Campo minado")
-                    .setFooter("jogo de " + nome());
-            rerender(hr.bundle());
+                    .setFooter("jogo de " + getNome());
+            rerender(hr.getBundle());
         }
 
         @Override
@@ -78,15 +72,16 @@ public class CriaCampoMinado extends ComandoGuild {
                 eb.addField(rb.getString("campo.state"),rb.getString(end ? "campo.over":"campo.clear"),false);
             else
                 eb.addField(rb.getString("campo.type"),check?"y":"x",false);
-            mensagem().editMessage(eb.build()).queue();
+            getMensagem().editMessageEmbeds(eb.build()).queue();
         }
 
         private void click(int n, Helper.Reacao r){
+            log.debug("click");
             if(check){
                 check=false;
                 end = campo.joga(x,n);
                 if(campo.finalizado()){
-                    rerender(r.bundle());
+                    rerender(r.getBundle());
                     log.info("end!!!");
                     finaliza();
                 }
@@ -94,6 +89,17 @@ public class CriaCampoMinado extends ComandoGuild {
             }
             x=n;
             check = true;
+        }
+
+        @Override
+        public String toString() {
+            return "ConversaCampo{" +
+                    "campo={\n\t" + campo +
+                    "\n}, eb=" + eb +
+                    ", x=" + x +
+                    ", check=" + check +
+                    ", end=" + end +
+                    '}';
         }
     }
 
