@@ -23,7 +23,7 @@ public class PollReactionRem extends ComandoGuildReacoes {
         final String tipo = args[0];
         PollHelper.getPollFromEmote(args,hm, dp -> {
             log.debug("callback");
-            dp.poll().remove(dp.index(), hm.id());
+            dp.poll().remove(dp.index(), hm.getId());
             PollHelper.reRender(hm,tipo,dp);
             Memoria.update(dp);
             return;
