@@ -1,8 +1,10 @@
 package rode.comando.guild.poll.texto;
 
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.Command;
-import net.dv8tion.jda.api.requests.restaction.CommandUpdateAction;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 import org.bson.Document;
 import rode.core.Anotacoes.EComandoPoll;
 import rode.core.Helper;
@@ -23,11 +25,11 @@ public class AdicionaOpcoesPoll extends AbrePoll {
     }
 
     @Override
-    public void subscribeSlash(CommandUpdateAction.CommandData commandData, ResourceBundle bundle) {
-        var subCommand = new CommandUpdateAction.SubcommandData(getCommand(), bundle.getString(getHelp()))
-                .addOption(new CommandUpdateAction.OptionData(Command.OptionType.STRING,"titulo","titulo da poll desejada").setRequired(true))
-                .addOption(new CommandUpdateAction.OptionData(Command.OptionType.STRING,"opcao","opção a ser adcional").setRequired(true));
-        commandData.addSubcommand(subCommand);
+    public void subscribeSlash(CommandData commandData, ResourceBundle bundle) {
+        var subCommand = new SubcommandData(getCommand(), bundle.getString(getHelp()))
+                .addOptions(new OptionData(OptionType.STRING,"titulo","titulo da poll desejada").setRequired(true),
+                        new OptionData(OptionType.STRING,"opcao","opção a ser adcional").setRequired(true));
+        commandData.addSubcommands(subCommand);
     }
 
     @Override
@@ -36,28 +38,28 @@ public class AdicionaOpcoesPoll extends AbrePoll {
             if(dp.guild() != null){
                 if(dp.opcoes().isEmpty()){
                     EmbedBuilder eb = Constantes.builder();
-                    eb.setTitle(hm.text("opcao.exec.empty"));
-                    help(eb,hm.bundle());
+                    eb.setTitle(hm.getText("opcao.exec.empty"));
+                    help(eb,hm.getBundle());
                     hm.reply(eb);
                     return;
                 }
                 for(String s:dp.opcoes())
                     if(Pattern.matches(".*<@!?\\d+>.*",s)){
-                        hm.reply(hm.text("opcao.exec.mention"));
+                        hm.reply(hm.getText("opcao.exec.mention"));
                         return;
                     }
                 Poll poll = dp.guild().getPoll(dp.titulo());
                 if(!poll.isOpen()){
-                    hm.replyTemp(hm.text("opcao.exec.close").formatted(poll.getTitle()));
+                    hm.replyTemp(String.format(hm.getText("opcao.exec.close"),poll.getTitulo()));
                     return;
                 }
 
                 poll.addOptions(dp.opcoes());
                 Memoria.guilds.updateOne(dp.query(),new Document("$set",dp.guild().toMongo()));
-                hm.reply(poll.makeDefaultEmbed(hm.bundle()), message->PollHelper.addReaction(message, poll.getOptions().size()));
+                hm.reply(poll.makeDefaultEmbed(hm.getBundle()), message->PollHelper.addReaction(message, poll.getOpcoes().size()));
                 return;
             }
-            hm.reply(hm.text("opcao.exec.404").formatted(dp.titulo()));
+            hm.reply(String.format(hm.getText("opcao.exec.404"),dp.titulo()));
         });
     }
 
