@@ -1,15 +1,16 @@
 package rode.comando.guild.poll.texto;
 
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.Command;
 import net.dv8tion.jda.api.entities.Guild;
-import net.dv8tion.jda.api.requests.restaction.CommandUpdateAction;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.ChartUtilities;
 import org.jfree.chart.JFreeChart;
 import org.jfree.data.general.DefaultPieDataset;
 import rode.core.Anotacoes.EComandoPoll;
-import rode.model.ComandoGuild;
 import rode.core.Helper;
 import rode.core.PollHelper;
 import rode.model.Poll;
@@ -17,21 +18,20 @@ import rode.model.Poll;
 import java.io.File;
 import java.io.IOException;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.ResourceBundle;
 
 @EComandoPoll
 public class GraficoPoll extends AbrePoll {
     public GraficoPoll() {
-        super("grafico", null, false,"grafico","grpah","gpoll");
+        super("grafico", null, false,"grafico","gráfico","graphic","gpoll");
         setPath("grafico");
     }
 
     @Override
-    public void subscribeSlash(CommandUpdateAction.CommandData commandData, ResourceBundle bundle) {
-        var subCommand = new CommandUpdateAction.SubcommandData(getCommand(),bundle.getString(getHelp()))
-                .addOption(new CommandUpdateAction.OptionData(Command.OptionType.STRING,"titulo","titulo da poll a ser deletada").setRequired(true));
-        commandData.addSubcommand(subCommand);
+    public void subscribeSlash(CommandData commandData, ResourceBundle bundle) {
+        var subCommand = new SubcommandData(getCommand(),bundle.getString(getHelp()))
+                .addOptions(new OptionData(OptionType.STRING,"titulo","titulo da poll a ser deletada").setRequired(true));
+        commandData.addSubcommands(subCommand);
     }
 
     @Override
@@ -43,7 +43,7 @@ public class GraficoPoll extends AbrePoll {
                 hm.reply(arq);
                 return;
             }
-            hm.reply(String.format(hm.text("grafico.exec"),dp.titulo()));
+            hm.reply(String.format(hm.getText("grafico.exec"),dp.titulo()));
         });
     }
 
@@ -61,7 +61,7 @@ public class GraficoPoll extends AbrePoll {
 
 
         JFreeChart chart = ChartFactory.createRingChart(
-                poll.getTitle(), pieDataset, false, false, false);
+                poll.getTitulo(), pieDataset, false, false, false);
 
         File f = new File("lixo/"+System.currentTimeMillis() + ".png");
         ChartUtilities.saveChartAsPNG(f, chart, 500, 350);
