@@ -1,17 +1,17 @@
 package rode.comando.guild.poll.texto;
 
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.Command;
-import net.dv8tion.jda.api.requests.restaction.CommandUpdateAction;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 import rode.core.Anotacoes.EComandoPoll;
-import rode.model.ComandoGuild;
 import rode.core.Helper;
 import rode.core.PollHelper;
 import rode.model.Poll;
 import rode.utilitarios.Constantes;
 
 import java.io.IOException;
-import java.util.LinkedList;
 import java.util.ResourceBundle;
 
 @EComandoPoll
@@ -22,10 +22,10 @@ public class MostraVotosPoll extends AbrePoll {
     }
 
     @Override
-    public void subscribeSlash(CommandUpdateAction.CommandData commandData, ResourceBundle bundle) {
-        var subCommand = new CommandUpdateAction.SubcommandData(getCommand(),bundle.getString(getHelp()))
-                .addOption(new CommandUpdateAction.OptionData(Command.OptionType.STRING,"titulo","titulo da poll desejada").setRequired(true));
-        commandData.addSubcommand(subCommand);
+    public void subscribeSlash(CommandData commandData, ResourceBundle bundle) {
+        var subCommand = new SubcommandData(getCommand(),bundle.getString(getHelp()))
+                .addOptions(new OptionData(OptionType.STRING,"titulo","titulo da poll desejada").setRequired(true));
+        commandData.addSubcommands(subCommand);
     }
 
     @Override
@@ -34,12 +34,12 @@ public class MostraVotosPoll extends AbrePoll {
             if(dp.guild() != null){
                 Poll poll = dp.guild().getPoll(dp.titulo());
                 EmbedBuilder eb = Constantes.builder();
-                eb.setTitle(String.format(hm.text("votos.exec.title"),dp.titulo()));
-                poll.getVotes(eb, hm.bundle());
+                eb.setTitle(String.format(hm.getText("votos.exec.title"),dp.titulo()));
+                poll.getVotes(eb, hm.getBundle());
                 hm.reply(eb);
                 return;
             }
-            hm.reply(String.format(hm.text("votos.exec.404"),dp.titulo()));
+            hm.reply(String.format(hm.getText("votos.exec.404"),dp.titulo()));
         });
     }
 }
