@@ -1,13 +1,14 @@
 package rode.comando.guild.poll.texto;
 
 import net.dv8tion.jda.api.Permission;
-import net.dv8tion.jda.api.entities.Command;
-import net.dv8tion.jda.api.requests.restaction.CommandUpdateAction;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 import org.bson.Document;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import rode.core.Anotacoes.EComandoPoll;
-import rode.model.ComandoGuild;
 import rode.core.Helper;
 import rode.core.PollHelper;
 import rode.model.Poll;
@@ -15,7 +16,6 @@ import rode.utilitarios.Constantes;
 import rode.utilitarios.Memoria;
 
 import java.io.IOException;
-import java.util.LinkedList;
 import java.util.ResourceBundle;
 
 @EComandoPoll
@@ -26,11 +26,11 @@ public class RemoveOpcoesPoll extends AbrePoll {
         setPath("remp");
     }
     @Override
-    public void subscribeSlash(CommandUpdateAction.CommandData commandData, ResourceBundle bundle) {
-        var subCommand = new CommandUpdateAction.SubcommandData(getCommand(), bundle.getString(getHelp()))
-                .addOption(new CommandUpdateAction.OptionData(Command.OptionType.STRING,"titulo","titulo da poll desejada").setRequired(true))
-                .addOption(new CommandUpdateAction.OptionData(Command.OptionType.STRING,"opção","opção a ser removida").setRequired(true));
-        commandData.addSubcommand(subCommand);
+    public void subscribeSlash(CommandData commandData, ResourceBundle bundle) {
+        var subCommand = new SubcommandData(getCommand(), bundle.getString(getHelp()))
+                .addOptions(new OptionData(OptionType.STRING,"titulo","titulo da poll desejada").setRequired(true),
+                        new OptionData(OptionType.STRING,"opção","opção a ser removida").setRequired(true));
+        commandData.addSubcommands(subCommand);
     }
 
     @Override
@@ -45,7 +45,7 @@ public class RemoveOpcoesPoll extends AbrePoll {
             log.debug("callback");
             if(dp.opcoes() == null || dp.opcoes().isEmpty()){
                 var eb = Constantes.builder();
-                help(eb,hm.bundle());
+                help(eb,hm.getBundle());
                 hm.reply(eb);
                 return;
             }
@@ -55,12 +55,12 @@ public class RemoveOpcoesPoll extends AbrePoll {
                 poll.remOptions(dp.opcoes());
                 Document d = dp.guild().toMongo();
                 Memoria.guilds.updateOne(dp.query(), new Document("$set",d));
-                hm.reply(poll.makeDefaultEmbed(hm.bundle()), message->
-                    PollHelper.addReaction(message,poll.getOptions().size())
+                hm.reply(poll.makeDefaultEmbed(hm.getBundle()), message->
+                    PollHelper.addReaction(message,poll.getOpcoes().size())
                 );
                 return;
             }
-            hm.reply(String.format(hm.text("remp.exec"),dp.titulo()));
+            hm.reply(String.format(hm.getText("remp.exec"),dp.titulo()));
         });
     }
 }
