@@ -1,5 +1,5 @@
 package rode.model.maker;
 
-public enum TipoLoop {
+public enum getTipo {
     G_MENSAGEM_TEXTO,G_MENSAGEM_REACAO;
 }
